some minor refactors
