+++ storagemarket/provider.go
@@ -18,7 +18,6 @@ import (
 	"github.com/filecoin-project/boost/db"
 	"github.com/filecoin-project/boost/filestore"
 	"github.com/filecoin-project/boost/storagemarket/types"
-	"github.com/filecoin-project/boost/storagemarket/types/dealcheckpoints"
 	"github.com/filecoin-project/go-address"
 	"github.com/filecoin-project/go-state-types/abi"
 	"github.com/filecoin-project/lotus/api/v1api"
@@ -27,7 +26,7 @@ import (
 	"github.com/libp2p/go-libp2p-core/event"
 )
 
-var log = logging.Logger("provider")
+var log = logging.Logger("boost-provider")
 
 type Config struct {
 	MaxTransferDuration time.Duration
@@ -81,7 +80,6 @@ func NewProvider(repoRoot string, dealsDB *db.DealsDB, fullnodeApi v1api.FullNod
 	}
 
 	return &Provider{
-		// TODO: pass max transfer duration as a param
 		config:    Config{MaxTransferDuration: 30 * time.Second},
 		Address:   addr,
 		newDealPS: newDealPS,
@@ -138,7 +136,7 @@ func (p *Provider) ExecuteDeal(dp *types.ClientDealParams) (pi *api.ProviderDeal
 	if err := p.validateDealProposal(ds); err != nil {
 		return &api.ProviderDealRejectionInfo{
 			Reason: fmt.Sprintf("failed validation: %s", err),
-		}, err
+		}, nil
 	}
 
 	// create a temp file where we will hold the deal data.
@@ -151,27 +149,16 @@ func (p *Provider) ExecuteDeal(dp *types.ClientDealParams) (pi *api.ProviderDeal
 		return nil, fmt.Errorf("failed to close temp file: %w", err)
 	}
 	ds.InboundFilePath = string(tmp.OsPath())
-
+	// make sure to remove the temp file if something goes wrong from here on.
 	defer func() {
 		if pi != nil || err != nil {
 			_ = os.Remove(ds.InboundFilePath)
 		}
 	}()
 
-	// send message to event loop to run the deal through the acceptance filter and reserve the required resources
-	// then wait for a response and return the response to the client.
-	respChan := make(chan acceptDealResp, 1)
-	select {
-	case p.acceptDealsChan <- acceptDealReq{rsp: respChan, deal: &ds}:
-	case <-p.ctx.Done():
-		return nil, p.ctx.Err()
-	}
-
-	var resp acceptDealResp
-	select {
-	case resp = <-respChan:
-	case <-p.ctx.Done():
-		return nil, p.ctx.Err()
+	resp, err := p.sendDealForAcceptance(&ds)
+	if err != nil {
+		return nil, fmt.Errorf("failed to send deal for acceptance: %w", err)
 	}
 
 	// if there was an error, we return no rejection reason as well.
@@ -184,7 +171,28 @@ func (p *Provider) ExecuteDeal(dp *types.ClientDealParams) (pi *api.ProviderDeal
 		return resp.ri, nil
 	}
 
+	log.Infow("scheduled deal for execution", "id", dp.DealUuid)
 	return nil, nil
+}
+
+func (p *Provider) sendDealForAcceptance(ds *types.ProviderDealState) (acceptDealResp, error) {
+	// send message to event loop to run the deal through the acceptance filter and reserve the required resources
+	// then wait for a response and return the response to the client.
+	respChan := make(chan acceptDealResp, 1)
+	select {
+	case p.acceptDealsChan <- acceptDealReq{rsp: respChan, deal: ds}:
+	case <-p.ctx.Done():
+		return acceptDealResp{}, p.ctx.Err()
+	}
+
+	var resp acceptDealResp
+	select {
+	case resp = <-respChan:
+	case <-p.ctx.Done():
+		return acceptDealResp{}, p.ctx.Err()
+	}
+
+	return resp, nil
 }
 
 func (p *Provider) Start(ctx context.Context) error {
@@ -195,25 +203,27 @@ func (p *Provider) Start(ctx context.Context) error {
 	// initialize the database
 	err := p.db.Init(p.ctx)
 	if err != nil {
-		return err
+		return fmt.Errorf("failed to init db: %w", err)
 	}
+	log.Infow("db initialized")
 
 	// restart all existing deals
 	deals, err := p.db.ListActive(p.ctx)
 	if err != nil {
-		return xerrors.Errorf("getting active deals: %w", err)
+		return fmt.Errorf("getting active deals: %w", err)
 	}
 
 	var restartWg sync.WaitGroup
 	for _, deal := range deals {
-		deal := deal
+		ds := deal
 		restartWg.Add(1)
 		go func() {
 			defer restartWg.Done()
 
 			select {
-			case p.restartDealsChan <- restartReq{deal: deal}:
+			case p.restartDealsChan <- restartReq{deal: ds}:
 			case <-p.ctx.Done():
+				log.Errorw("failed to restart deal", "err", p.ctx.Err(), "id", ds.DealUuid)
 			}
 		}()
 	}
@@ -262,101 +272,4 @@ func (p *Provider) CancelDeal(ctx context.Context, dealUuid uuid.UUID) error {
 	}
 	dh.cancel(ctx)
 	return nil
-}
-
-type acceptDealReq struct {
-	rsp  chan acceptDealResp
-	deal *types.ProviderDealState
-}
-
-type acceptDealResp struct {
-	accepted bool
-	ri       *api.ProviderDealRejectionInfo
-	err      error
-}
-
-type failedDealReq struct {
-	st  *types.ProviderDealState
-	err error
-}
-
-type restartReq struct {
-	deal *types.ProviderDealState
-}
-
-// TODO: This is transient -> If it dosen't work out, we will use locks.
-// 1:N will move this problem elsewhere.
-func (p *Provider) loop() {
-	defer p.wg.Done()
-
-	for {
-		select {
-		case restartReq := <-p.restartDealsChan:
-			log.Infow("restart deal", "id", restartReq.deal.DealUuid)
-
-			// Put ANY RESTART SYNCHRONIZATION LOGIC HERE.
-			// ....
-			//
-			p.wg.Add(1)
-			go func() {
-				defer p.wg.Done()
-
-				p.doDeal(restartReq.deal)
-			}()
-
-		case dealReq := <-p.acceptDealsChan:
-			deal := dealReq.deal
-			log.Infow("process accept deal request", "id", deal.DealUuid)
-
-			writeDealResp := func(accepted bool, ri *api.ProviderDealRejectionInfo, err error) {
-				select {
-				case dealReq.rsp <- acceptDealResp{accepted, ri, err}:
-				case <-p.ctx.Done():
-					return
-				}
-			}
-
-			var err error
-			if err != nil {
-				go writeDealResp(false, nil, err)
-				continue
-			}
-
-			// TODO: Deal filter, storage space manager, fund manager etc . basically synchronization
-			// send rejection if deal is not accepted by the above filters
-			accepted := true
-			if !accepted {
-				go writeDealResp(false, &api.ProviderDealRejectionInfo{}, nil)
-				continue
-			}
-			go writeDealResp(true, nil, nil)
-
-			// write deal state to the database
-			log.Infow("insert deal into DB", "id", deal.DealUuid)
-
-			deal.CreatedAt = time.Now()
-			deal.Checkpoint = dealcheckpoints.New
-
-			err = p.db.Insert(p.ctx, deal)
-			if err != nil {
-				go writeDealResp(false, nil, err)
-				continue
-			}
-
-			// start executing the deal
-			p.wg.Add(1)
-			go func() {
-				defer p.wg.Done()
-
-				p.doDeal(deal)
-			}()
-
-		case failedDeal := <-p.failedDealsChan:
-			log.Errorw("deal failed", "id", failedDeal.st.DealUuid, "err", failedDeal.err)
-			// Release storage space , funds, shared resources etc etc.
-
-		case <-p.ctx.Done():
-			return
-		}
-	}
 }
